requirements.txt
streamlit
tensorflow
numpy
pandas
sklearn
flask
opencv-python
os
cv2
pyngrok
pillow